add libc functions, remove floating point support, can now compile itself

+++ test/offsetof.c
@@ -5,7 +5,7 @@ typedef struct {
   int a;
   char b;
   int c;
-  double d;
+  long long d;
 } T;
 
 int test_main() {
